+++ app/models/document.rb
@@ -1,0 +1,4 @@
+class Document < ApplicationRecord
+  has_rich_text :body
+  has_one :resource, as: :resourceable, dependent: :destroy
+end